+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Minor bug fixes in STDIO server
 - More efficient client reads
 - Rename server methods `prepare` for clarity
+- Improved socket state handling
 
 ## 0.1.0 - December 20, 2018
 First release